update video.js and fix pale moon-only issue
The pale moon issue was caused by the data-setup parameter that caused the warning below:

VIDEOJS: WARN: Player "sb-video" is already initialised. Options will not be applied.

The data-setup parameter appears to be unused and empty, but Pale Moon uses it instead of the player variable, making the video not autoplay and making the control bar not fully wide if the video is not 16:9. Removing the data-setup parameter gives Pale Moon the intended behavior seen on other browsers like Internet Explorer.