stupid comment implementation on new qobo (also add ratelimit)
the way commenting worked on opensb finalium is that it needs a bunch of post
variables like "really" and "type". "really" is just to actually confirm if the
request is valid at the slightest, while "type" is supposed to differentiate
between submission comments and user profile comments, which are in two
different tables since I didn't care about db tables back in 2021.

the comment form on new qobo is not done with javascript, it's done with
html, which is why I called this stupid. we'll plan on adding javascript,
but it's probably going to be a garbage mess of jquery and vanilla js mixed
together for no reason.