remove email hashing from privacy policy
the problem with email hashing is that not every single qobo account has a proper email (we're not talking about the ones with the "go@fuck.yourself" kinds of emails). most accounts had a generic squarebracket email address embedded onto it, and they're nearly all abandoned.

we plan on having some sort of email verification thing in the future, and hashing emails would make that nearly impossible.