Fix 999.999.999.999 IP bug and add user banning into site
Decided to implement in-site user banning since 2 days ago we got a porn-bot and someone was trying to impersonate me under "Bluffingo2".

We also hide banned users off the user list page, and we now hide the ID as on production there's a pretty jarring jump between 205 and 1654 due to a register-spamming incident in June 2023.