split compile-scss back into 2 because the OS detection code is broken on bash 5.1 which is what the production server uses and i can't be bothered to fight with debian's tendency to stick with outdated packages because "stability".
opensb is developed on microsoft windows. "git on windows" bundles bash 5.2 as of writing this. these linux-specific issues are always an oversight. the production server uses debian 11 and i'm pretty sure updating it to debian 12 has to be done with some control panel page within whatever VPS service sks2002 uses for the qobo vps, and it might break shit.